+++ packages/session-replay-browser/src/hooks/click.ts
@@ -5,6 +5,8 @@ import { SessionReplayEventsManager as AmplitudeSessionReplayEventsManager } fro
 import { PayloadBatcher } from 'src/track-destination';
 import { finder } from '../libs/finder';
 import { getGlobalScope, ILogger } from '@amplitude/analytics-core';
+import { UGCFilterRule } from 'src/config/types';
+import { getPageUrl } from '../helpers';
 
 // exported for testing
 export type ClickEvent = {
@@ -26,6 +28,7 @@ type Options = {
   deviceIdFn: () => string | undefined;
   eventsManager: AmplitudeSessionReplayEventsManager<'interaction', string>;
   mirror: Mirror;
+  ugcFilterRules: UGCFilterRule[];
 };
 
 const HOUR_IN_MILLISECONDS = 3_600_000;
@@ -70,7 +73,7 @@ export const clickBatcher: PayloadBatcher = ({ version, events }) => {
 };
 
 export const clickHook: (logger: ILogger, options: Options) => mouseInteractionCallBack =
-  (logger, { eventsManager, sessionId, deviceIdFn, mirror }) =>
+  (logger, { eventsManager, sessionId, deviceIdFn, mirror, ugcFilterRules }) =>
   (e) => {
     if (e.type !== MouseInteractions.Click) {
       return;
@@ -104,6 +107,8 @@ export const clickHook: (logger: ILogger, options: Options) => mouseInteractionC
 
     const { left: scrollX, top: scrollY } = getWindowScroll(globalScope as unknown as Window);
 
+    const pageUrl = getPageUrl(location.href, ugcFilterRules);
+
     const event: ClickEvent = {
       x: x + scrollX,
       y: y + scrollY,
@@ -111,7 +116,7 @@ export const clickHook: (logger: ILogger, options: Options) => mouseInteractionC
 
       viewportHeight: innerHeight,
       viewportWidth: innerWidth,
-      pageUrl: location.href,
+      pageUrl,
       timestamp: Date.now(),
       type: 'click',
     };