+++ packages/plugin-autocapture-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.9](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-autocapture-browser@1.2.8...@amplitude/plugin-autocapture-browser@1.2.9) (2025-06-11)
+
+**Note:** Version bump only for package @amplitude/plugin-autocapture-browser
+
+
+
+
+
 ## [1.2.8](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-autocapture-browser@1.2.7...@amplitude/plugin-autocapture-browser@1.2.8) (2025-06-03)
 
 **Note:** Version bump only for package @amplitude/plugin-autocapture-browser